% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lvm.r
\name{generate.ln}
\alias{generate.ln}
\title{method for Polytomous Response: Response Time (Lognormal)}
\usage{
\method{generate}{ln}(info)
}
\description{
method for Polytomous Response: Response Time (Lognormal)
}
