% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_generate_lvm.r
\name{generateLV.ln}
\alias{generateLV.ln}
\title{method for Polytomous Response: Response Time (Lognormal)}
\usage{
\method{generateLV}{ln}(info)
}
\description{
method for Polytomous Response: Response Time (Lognormal)
}
