% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rstan.r
\name{loadRstan}
\alias{loadRstan}
\title{Load rstan model}
\usage{
loadRstan(lv_type = "2pl")
}
\arguments{
\item{lv_model}{a character specifying a latent model}
}
\value{
An object of class \code{\linkS4class{stanmodel}}
}
\description{
Load rstan model
}
\examples{
stan_model <- rstan_path(lv_type = "rasch")
}
