% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.r
\name{makeFLPSdata}
\alias{makeFLPSdata}
\title{#' S3 generic for data setting depending on latent variable models
#'
dataSetting <- function(info, ...) {
  UseMethod("dataSetting", info)
}}
\usage{
makeFLPSdata(inp_data, outcome, group, covariate, lv_model, lv_type)
}
\description{

}
\details{

}
