% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.r
\name{makeInpData}
\alias{makeInpData}
\title{make dataframe style data for simulation}
\usage{
makeInpData(N, R2Y, omega, tau0, tau1, lambda, R2eta, nsec, lvmodel)
}
\description{
make dataframe style data for simulation
}
