% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.r
\name{my.gelman.diag}
\alias{my.gelman.diag}
\title{gelman diagnostic with chol.default error addressed.}
\usage{
my.gelman.diag(
  x,
  confidence = 0.95,
  transform = FALSE,
  autoburnin = FALSE,
  multivariate = TRUE
)
}
\description{
gelman diagnostic with chol.default error addressed.
}
