% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FLPS.r
\name{runFLPS}
\alias{runFLPS}
\title{Conduct fully latent principal stratification}
\usage{
runFLPS(
  inp_data = NULL,
  flps_data = NULL,
  outcome,
  group,
  covariate,
  lv_model,
  lv_type,
  stan_options = list(...),
  ...
)
}
\arguments{
\item{inp_data}{a dataframe}

\item{flps_data}{a list}
}
\description{
Conduct fully latent principal stratification
}
\examples{

}
