% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.r
\name{makeDat}
\alias{makeDat}
\title{Generate Fully Latent Principal Stratification data}
\usage{
makeDat(N, R2Y, omega, tau0, tau1, lambda, R2eta, nsec, lvmodel)
}
\arguments{
\item{N}{a numeric indicating sample size.}

\item{R2Y}{a numeric indicating predictive power of covariates.}

\item{omega}{a numeric indicating the relationship between eta_T and Y_C.}

\item{tau0}{a numeric indicating the treatment eff when eta_T=0.}

\item{tau1}{a numeric indicating the relationship between eta_T and Y_T-Y_C; (for now, single level (not multilevel), linear, normal).}

\item{lambda}{a numeric indicating the mean of Worked problems/person.}

\item{R2eta}{a numeric indicating Predictive power of latent variable (extent to which covariates predict eta).}

\item{nsec}{a numeric indicating the number of maximum sections given to students.}

\item{lvmodel}{a character specifying a type of latent variable model.}
}
\value{
a list containing the data for running FLPS.
}
\description{
\code{\link{makeDat}} is a function for generating a data based on the given
information.
}
\examples{
sdat <- makeDat(
  N = 100,
  R2Y = 0.2,
  omega = 0.2,
  tau0 = 0.13,
  tau1 = -0.06,
  lambda = 10,
  R2eta = 0.5,
  nsec = 10,
  lvmodel = "sem"
)

}
