% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FLPS.r
\name{runFLPS}
\alias{runFLPS}
\title{Conduct fully latent principal stratification}
\usage{
runFLPS(
  inp_data = NULL,
  flps_data = NULL,
  outcome,
  group,
  covariate,
  lv_model,
  lv_type,
  stan_options = list(...),
  ...
)
}
\arguments{
\item{inp_data}{A matrix}

\item{flps_data}{A list}

\item{outcome}{A character indicating the name of an outcome variable}

\item{group}{A character indicating the name of a treatment/control group variable}

\item{covariate}{A character indicating the names of covariates variables}

\item{lv_model}{A description of the latent variable model, which is similar to the lavaan model syntax.}

\item{lv_type}{A character indicating the type of latent variable models}

\item{stan_options}{A list containing \code{\link{stan}} options, using 'name = value'.}

\item{...}{Additional options.}
}
\description{
Conduct fully latent principal stratification
}
\examples{

}
