% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.r
\name{runSimulation}
\alias{runSimulation}
\title{Run FLPS Simulation}
\usage{
runSimulation(
  param_list,
  iter = 2000,
  warmup = 1000,
  cores = 1,
  chains = 1,
  pars = NA
)
}
\arguments{
\item{param_list}{a list of parameters}

\item{iter}{deault 2000}

\item{warmup}{default 1000}

\item{cores}{default 1}

\item{chains}{default 1}
}
\value{
A list containing the parameter list, the simulated data and the stan fit
}
\description{
Simulation of fully latent principal stratification
}
\details{
test a simulation.
}
\examples{
pars <- list(
N = 500,
R2Y = 0.2,
omega = 0.2,
tau0 = 0.13,
tau1 = -0.06,
lambda = 10,
R2eta = 0.2,
nsec = 20,
lvmodel = 'rasch'
)
\donttest{
runSimulation(pars)
}
}
